requirements.txt
ccxt==4.4.26
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.155.0
google-auth-oauthlib==1.2.1
python-json-logger==2.0.7
